+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 * `fetch.download_file` now uses content disposition header when available to determine filenames
+* The `get_orb` module now leverages the Copernicus Sentinels POD Data Hub API instead of the deprecated
+  qc.sentinel1.eo.esa.int API
 
 ## [1.6.6](https://github.com/ASFHyP3/hyp3-lib/compare/v1.6.5...v1.6.6)